Add receiver builder
The first typestate has always been an odd case. Stricly speaking you
cannot resume from the uninitialized state. The earliest state you can
resume from is `Initialized`. So this commit removes UnInitialized as a
"state" and replaces it with a builder.

Additionally a builder model allows the application to access
the short id after the builder is created. The application may use
session id as the key for the perister which is needed when you build().